requirements.txt
pillow~=10.4
